requirements.txt
faiss-gpu==1.7.2
farm-haystack==1.4.0
nltk==3.7
numba==0.55.2
numpy==1.21.6
openpyxl==3.0.10
pandas==1.3.5
scikit-learn==1.0.2
scipy==1.7.3
sentence-transformers==2.2.0
sentencepiece==0.1.96
tokenizers==0.10.3
transformers==4.13.0